+++ api/src/main/java/org/openmrs/module/tag/Tag.java
@@ -7,7 +7,7 @@
  * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
  * graphic logo is a trademark of OpenMRS Inc.
  */
-package org.openmrs;
+package org.openmrs.module.tag;
 
 import org.openmrs.BaseOpenmrsData;
 import org.openmrs.User;
