+++ omod/src/main/resources/config.xml
@@ -12,7 +12,7 @@
 			${project.parent.description}
 	</description>
 
-	<activator>org.openmrs.TagActivator</activator>
+	<activator>org.openmrs.module.tag.TagActivator</activator>
 	
 	<!-- /Base Module Properties -->
